+++ src/main/java/net/mobz/data/Loots.java
@@ -12,6 +12,7 @@ import net.minecraft.world.level.storage.loot.parameters.LootContextParamSets;
 public class Loots {
 	public static LootTableProvider all(PackOutput packOutput, CompletableFuture<HolderLookup.Provider> lookupProvider) {
 		return new LootTableProvider(packOutput, Set.of(), List.of(
+				new LootTableProvider.SubProviderEntry(RewardLoot::new, LootContextParamSets.GIFT),
 				new LootTableProvider.SubProviderEntry(BlockLoot::new, LootContextParamSets.BLOCK),
 				new LootTableProvider.SubProviderEntry(EntityLoot::new, LootContextParamSets.ENTITY)
 			), lookupProvider);